Public Documentation

Documentation for PLaplace.jl's public interface.

See the Internal page of the library for the documentation of internal types and functions.

Contents

Public Documentation

Contents

Module

Algorithm

Keyword Types

Initial prolongations

FEM

Modified Derivatives

Statistics

PLaplace Data

Log Data

Error Analysis

Other Utilities

Module

PLaplace.PLaplace — Module

PLaplace

A Solver for Problems of the p-Laplacian PDE Operator.

This package imports the following packages:

Base

Core

DocStringExtensions

IncompleteLU

InvertedIndices

IterativeSolvers

LinearAlgebra

MinFEM

Preconditioners

Printf

SparseArrays

WriteVTK

Algorithm

PLaplace.PLaplaceData — Type

mutable struct PLaplaceData

Main output type for solution of a p-Laplace problem. Stores all parameters and results obtained by the solution of a pLaplace problem. Can be used for Statistics, Error Analysis and writing the solution to a .vtk-file.

Fields

mesh::Mesh: Mesh of computational domain.

dirichlet_nodes::Set{Int64}: Dirichlet boundary of the computational domain.

fixed_nodes::Union{Missing, Set{Int64}}: Fixed nodes on the boundary of the computational domain.

neumann_elements::Union{Missing, Set{Int64}}: Neumann boundary of the computational domain.

f::Union{Missing, AbstractVector{Float64}}: Domain source term.

h::Union{Missing, AbstractVector{Float64}}: Neumann boundary condition.

qdim::Int64: Number of components of the solution.

p::Float64: PDE parameter.

gradient_bound::Union{Missing, Float64}: Upper bound for the gradient of the solution.

eps::Float64: Accuracy of the solution in a variational sense.

stepsize::Stepsize: Stepping scheme of the interior point method.

Naux::Union{Missing, Int64}: Number of iterations in the auxilliary path-following. Is missing if algorithm did not reach auxilliary stage.

Nmain::Union{Missing, Int64}: Number of iterations for the main path-following Is missing if algorithm did not reach main stage.

tsetup::Union{Missing, Float64}: Time required for the setup Is missing if algorithm did not reach setup stage.

taux::Union{Missing, Float64}: Time required for the auxilliary path-following Is missing if algorithm did not reach auxilliary stage.

tmain::Union{Missing, Float64}: Time required for the main path-following Is missing if algorithm did not reach main stage.

calculation_nodes::Union{Missing, Set{Int64}}: Nodes on which the solution was computed, i.e. nodes without Dirichlet condtion. Is missing if algorithm failed during setup stage.

g::Union{Missing, AbstractVector{Float64}}: Dirichlet boundary condition prolonged to the whole domain. Is missing if algorithm failed during setup stage.

u::Union{Missing, AbstractVector{Float64}}: Shifted solution on calculation nodes without prolonged boundary condition. Is missing if algorithm did not converge.

v::Union{Missing, AbstractVector{Float64}}: Solution to the p-Laplace problem. Is missing if algorithm did not converge.

msg::String: Notifications from the iteration. In particular contains changes of solvers and preconditioners and information on early stops.

PLaplace.solve_plaplace — Function

solve_plaplace(
    p::Float64,
    mesh::Mesh,
    g::Union{AbstractVector{Float64}, Function},
    dirichlet_boundary::Set{Boundary};
    <keyword arguments>
) -> PLaplaceData

Returns solution to a given problem for the p-Laplace operator with a Dirichlet condition on (a part of) the boundary. Additionally, a volume source term and Neumann conditions on a non-intersecting part of the boudary can be specified. The solution is computed by an interior-point method applied to a reformulated problem. For more information on the algorithm, please look at the algorithm page of the documentation.

Below you find a list of mandatory and keyword arguments indicating which functions can be specified continously or discrete and which parameters can be adjusted. Further, the outputs can be controlled in particular if log and vtk files are generated. It is also possible to track the condition number of the system matrix during the iteration, but please not that this significantly impacts the performance and requires a lot of memory.

Mandatory Arguments

p::Float64: problem parameter.

mesh::Mesh: FEM mesh of the domain.

g::Union{AbstractVector{Float64}, Function}: Dirichlet boundary condition.

dirichlet_boundary::Union{Set{Boundary}, Set{Int64}}: Parts of the physical boundary to hold the Dirichlet condition. Either given by set of named boundaries or node ids.

Keyword Arguments

f::Union{AbstractVector{Float64}, Function, Missing} = missing: Discrete or continous source term.

neumann_boundary::Union{Set{Boundary}, Set{Int64}, Missing} = missing: Parts of the physical boundary to hold Neumann condition. Either given by set of named boundaries or boundary element ids.

h::Union{AbstractVector{Float64}, Function, Missing} = missing: Discrete or continous Neumann boundary condition. Requires neumann_boundary to be set.

boundary_prolongation::Union{AbstractVector{Float64}, Missing} = missing: Prescribed discrete prolongation of the Dirichlet condition to the full domain. Will be computed during the algorithm if not set.

fixed_nodes::Union{Set{Boundary}, Set{Int64}, Missing} = missing,: Set of fixed node indices, i.e. additional homogeneous Dirichlet nodes. Either given by set of named boundaries or node ids. Is ignored if boundary_prolongation is set.

qdim::Int64 = 1: Number of components of the problem, respectively f, g and h as well as u and v. Is mandatory to set for vector-valued setting.

eps::Float64 = 1e-6: Accuracy of the solution. In particular termination criterion for the main path-following.

stepsize::Stepsize = ADAPTIVE: Specifier for using an alternative stepping scheme in the internal path-following.

maxiterations::Int64 = 1000: Maximal number of iteriations per stage of the internal path-following schemes.

kappa::Float64 = 10.0: Parameter used for updates of the step length in long-step path-following variants.

backtracking_maxiterations::Int64 = 25: Maximal number of iterations in backtracking search for the update of the iterate in long-step path-following variants.

backtracking_decrementfactor::Float64 = 0.25: Decrement factor in backtracking search of long-step path-following variants.

solver::LinearSolver = CHOLESKY: Specifier for the solver internally used for linear systems.

preconditioner::Preconditioner = NONE: Specifier for the preconditioner internally used for linear systems.

useharmonicprolongation::Bool = true: Flag for prolongation of boundary condition as solution to the harmonic problem or by 0. Is ignored if boundary_prolongation is set.

vtkfile::Union{String,Missing} = missing: File name for writing solution to vtk. If not set, no file will be written.

verbose::Bool = true: Flag if log is written to console during the iteration.

logfile::Union{String,Missing} = missing: File name for writing log to a file. If not set, log will not be exported.

logcondition::Bool = false: Flag for tracking condition of the system matrix is tracked and included in log file. Will only be exported if logfile is set.

Keyword Types

PLaplace.Stepsize — Type

primitive type Stepsize <: Enum{Int32} 32

Public type used to specify the stepping scheme of the path-following schemes. For more information see the interior-point section in the documentation.

Available Options

SHORT: Regular update of the paramter t.

LONG: Larger update of the parameter t if iterate fulfill approximate centering conditon. Results in line-search for application of update on the iterate x.

ADAPTIVE: Same as the long stepping, but the factor for the larger update is adaptively changed depending on how many steps it required to fulfill the approximate centering condition.

PLaplace.LinearSolver — Type

primitive type LinearSolver <: Enum{Int32} 32

Public type used to specify the linear solver in the solution algorithm.

Available Options

LU: Direct solver with LU decomposition.

CHOLESKY: Direct solver with Cholesky decomposition. Default options since Hessian is supposed to be symmetric positive definite.

CG: Iterative solver using the conjugate gradients methods.

BICGSTAB: Iterative solver using the bi-conjugate gradients stabilized method.

GMRES: Iterative solver using the generalized minimal residual method. Has proven convergence for unsymmetric systems, but requires a lot of memory.

PLaplace.Preconditioner — Type

primitive type Preconditioner <: Enum{Int32} 32

Public type used to specify a preconditioner for the linear systems in the solution process.

Available Options

NONE: Dummy option for no preconditioner. In particular used with direct solvers.

JACOBI: Preconditioner with diagonal matrix.

ILU: Preconditioner with incomplete LU decomposition.

ICHOLESKY: Preconditioner with incomplete Cholesky decomposition.

AMG: Preconditioner with algebraic multi-grid.

Initial prolongations

PLaplace.compute_prolongation_harmonic — Function

compute_prolongation_harmonic(
    g::AbstractVector{Float64},
    mesh::Mesh,
    has_sourceterm::Bool,
    f::AbstractVector{Float64}, 
    dirichlet_nodes::Set{Int64},
    has_neumannboundary::Bool,
    h::AbstractVector{Float64},
    neumann_elements::Set{Int64}, 
    solveLS::Function,
    preconditioner::Function;
    qdim::Int64 = 1
) -> AbstractVector{Float64}

Returns discrete prolongation of the boundary condition g to the whole domain as solution of the corresponding linear Laplace problem, i.e. $p = 2$.

Mandatory Arguments

g::AbstractVector{Float64}: Discretely evaluated boundary condition.

mesh::Mesh: Mesh of the domain.

has_sourceterm::Bool: Flag for evaluating source term.

f::AbstractVector{Float64}: Source term.

dirichlet_nodes::Set{Int64}: Nodes of the boundary to hold Dirichlet condition.

has_neumannboundary::Bool: Flag for evaluating Neumann boundary.

h::AbstractVector{Float64}: Neumann boundary condition.

neumann_elements::Set{Int64}: Edges of the boundary to hold Neumann condition.

solveLS::Function: Function pointer to solve linear system.

preconditioner::Function: Function pointer to compute preconditioner for linear system.

Keyword Arguments

qdim::Int64 = 1: Number of components of f, h and g.

PLaplace.compute_prolongation_zero — Function

compute_prolongation_zero(
    g::AbstractVector{Float64},
    mesh::Mesh;
    qdim::Int64 = 1
) -> AbstractVector{Float64}

Returns discrete prolongation of g to the whole domain by zero on every node.

Returns discrete prolongation of the boundary condition g to the whole domain by zero on every node.

Mandatory Arguments

g::AbstractVector{Float64}: Discretely evaluated boundary condition.

mesh::Mesh: Mesh of the domain.

Keyword Arguments

qdim::Int64 = 1: Number of components of f, h and g.

FEM

MinFEM.Mesh — Type

mutable struct Mesh

Type for a triangular finite element mesh with volume and boundary markers.

Fields

d::Int64: Spatial dimension

nnodes::Int64: Number of nodes

nelems::Int64: Number of elements

nboundelems::Int64: Number of physical (marked) boundary elements

Nodes::Vector{Vector{Float64}}: List of node coordinates

Elements::Vector{Vector{Int64}}: List of element node indices

BoundaryElements::Vector{Vector{Int64}}: List of boundary element node indices

ParentElements::Vector{Int64}: List of parent elements to boundary elements

ParentBoundaries::Vector{Int64}: List of parent element local boundary to boundary elements

Boundaries::Dict{Int64, Boundary}: Dictionary of marked boundaries

Domains::Dict{Int64, Domain}: Dictionary of marked volume regions

Entities::Vector{Dict{Int64, MinFEM.Entity}}: Dictionary of physical entities

MinFEM.Boundary — Type

mutable struct Boundary <: MinFEM.Region

Structure holding the name and sets of node and edge indices for one particular physical boundary.

Fields

Name::String: Unique physical name

Nodes::Set{Int64}: List of node indices

Elements::Set{Int64}: List of element indices

MinFEM.import_mesh — Function

import_mesh(fileName::String) -> Mesh

Returns a mesh imported from a gmsh file of version v1, v2 or v4.

MinFEM.select_boundaries — Function

select_boundaries(mesh::Mesh, args...) -> Set{Boundary}

Returns set of all or specified physical boundaries of the mesh.

Modified Derivatives

PLaplace.assemble_derivativetensor — Function

assemble_derivativetensor(
    mesh::Mesh;
    qdim::Int64 = 1
) -> Dict{Tuple{Int64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}}

Returns the discrete derivative tensor for all elements of mesh and the number of components of the function qdim. Each key tuple (j,r) represents the derivative matrix for the component r in direction x_j.

PLaplace.assemble_derivativetensor_boundary — Function

assemble_derivativetensor_boundary(
    mesh::Mesh,
    BoundaryElements::Set{Int64}; 
    qdim::Int64 = 1
) -> Dict{Tuple{Int64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}}

Returns the discrete derivative tensor for all or specified boundary elements of mesh and number of components qdim. Each key tuple (j,r) represents the derivative matrix for the component r in direction x_j. Workaround based on the derivate tensor for the corresponding full element and the fact that the gradient is constant on the element.

PLaplace.assemble_derviativetensor_modified — Function

assemble_derviativetensor_modified(
    D::AbstractDict{Tuple{Int64,Int64},SparseMatrixCSC{Float64,Int64}},
    nodesToDrop::Set{Int64};
    qdim::Int64 = 1
) -> Dict{Tuple{Int64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}}

Returns a new derivative tensor reduced by nodesToDrop, e.g used to drop homogeneous Dirichtlet boundary points of the mesh.

PLaplace.compute_derivative — Function

compute_derivative(
    D::AbstractDict{Tuple{Int64,Int64},SparseMatrixCSC{Float64,Int64}},
    v::AbstractVector{Float64}
) -> Dict{Tuple{Int64, Int64}, AbstractVector{Float64}}

Returs coefficient vectors of all the first partial derivatives $D^{(j,r)}v$ of a function given as coefficient vector v using the discrete derivative tensor D. Each key tuple (j,r) represents the derivative of the component r in direction x_j.

PLaplace.compute_normalderivative — Function

compute_normalderivative(
    mesh::Mesh,
    boundary::Set{Boundary},
    v::AbstractVector{Float64};
    qdim::Int64 = 1
) -> Dict{Tuple{Int64, Int64}, AbstractVector{Float64}}

Returns normal derivative of a function on all or specified boundary elements of the mesh. Each key r represents the derivative of the component r in outer normal direction.

compute_normalderivative(
    mesh::Mesh,
    boundary::Set{Boundary},
    v::AbstractVector{Float64};
    qdim::Int64 = 1
) -> Dict{Tuple{Int64, Int64}, AbstractVector{Float64}}

Same as previous compute_normalderivative(...). However takes Set{Boundary} of named boundary objects as argument boundary and translates it to a Set{Int64} containing the indices of the nodes contained in the given boundaries.

Statistics

PLaplace Data

PLaplace.hasresult — Function

hasresult(data::PLaplaceData) -> Bool

Returns if the object contains a result, i.e. if an algorithm ran and converged. Should be called before the result is used via data.v.

PLaplace.write_result_to_vtk — Function

write_result_to_vtk(
    filename::String,
    data::PLaplaceData
) -> Union{Nothing, Vector{String}}

Writes result vector to a VTK file with the given name. The file itself will be created, but the path has to exist prior.

PLaplace.write_result_to_txt — Function

write_result_to_txt(filename::String, data::PLaplaceData)

Writes result vector to a VTK file with the given name. The file itself will be created, but the path has to exist prior.

PLaplace.print_defaultdata — Function

print_defaultdata(data::PLaplaceData) -> Any

Prints problem data that had to be specified for the algorithm ot the console.

PLaplace.print_statistics — Function

print_statistics(data::PLaplaceData) -> Any

Prints statistics of the algorithm run ot the console.

PLaplace.write_statistics_header — Function

write_statistics_header(filename::String; guarded::Bool=false)

Clears and writes a header for a statistics log to the given file. In case the file does not exist, it will be created, but only if the path exists. If guarded checks before if file already contains a header and then does not overwrite potentially previous results.

PLaplace.check_statistics_header — Function

check_statistics_header(filename::String) -> Bool

Checks if given file exists and alredy contains a statistics header.

PLaplace.write_statistics — Function

write_statistics(
    filename::String,
    data::PLaplaceData
) -> Int64

Writes statistics line corresponding to the header to a given log file.

Log Data

PLaplace.AlgorithmLogData — Type

mutable struct AlgorithmLogData

Stores data read from an algorithm log file. Can be used to generate plots from one solution process.

Fields

totalIterations::Vector{Union{Missing, Int64}}: Running number of total iterations for each iteration.

sections::Vector{String}: Section tag in each iteration.

partIterations::Vector{Union{Missing, Int64}}: Running number of iterations in current section for each iteration.

types::Vector{Union{Missing, String}}: Tag for type of the update step for each iteration. Only relevant for long and adaptive stepping schemes.

lastaccepts::Vector{Union{Missing, Int64}}: Number of iterations since previous accept for each iteration Only relevant for adaptive stepping scheme.

kappas::Vector{Union{Missing, Float64}}: Parameter kappa for each iteration. Only relevant for adaptive stepping scheme.

accnorms::Vector{Union{Missing, Float64}}: Norm value for acceptance of step for each iteration. Only relevant for long and adaptive stepping schemes.

searchIterations::Vector{Union{Missing, Int64}}: Number of iterations in backtracking linesearch for each iteration. Only relevant for long and adaptive stepping schemes

scalings::Vector{Union{Missing, Float64}}: Scaling of the update generated by the backtracking for each iteration. Only relevant for long and adaptive stepping schemes.

ts::Vector{Union{Missing, Float64}}: Value of t for each iteration.

criteria::Vector{Union{Missing, Float64}}: Value of convergence criterum for each iteration.

bounds::Vector{Union{Missing, Float64}}: Bound for convergence criteria for each iteration. Is constant per section.

conditions::Vector{Union{Missing, Float64}}: Condition number of the system matrix for each iteration.

PLaplace.read_algorithmlog — Function

read_algorithmlog(filename::String)

Returns AlgorithmLogData by reading the given algorithm log file.

Error Analysis

PLaplace.ErrorData — Type

mutable struct ErrorData

Structure to hold error values compared to analytical solutions for multiple computations. Can be used to generate plots.

Fields

p::Vector{Float64}: PDE parameters.

n::Vector{Int64}: Number of gridpoints

eps::Vector{Float64}: Accuracy

errors::Vector{Vector{Float64}}: Errors

PLaplace.objective_functional — Function

objective_functional(
    u::AbstractVector{Float64},
    p::Float64,
    mesh::Mesh; 
    f::Union{AbstractVector{Float64}, Function, Missing} = missing,
    neumann_boundary::Union{Set{Boundary}, Set{Int64}, Missing} = missing,
    h::Union{AbstractVector{Float64}, Function, Missing} = missing,
    qdim::Int64 = 1
) -> Float64

Returns value of variational formulation functional for the p-Laplace problem evaluated at u.

PLaplace.compute_errors — Function

compute_errors(
    anasol::Vector{Float64},
    numsol::Vector{Float64},
    mesh::Mesh,
    neumann_boundary::Union{Missing, Set{Int64}, Set{Boundary}},
    h::Union{Missing, AbstractVector{Float64}},
    f::Union{Missing, AbstractVector{Float64}},
    p::Float64,
    qdim::Int64
) -> Vector{Float64}

Returns various errors of numerical solution compared to given discretized analytical solution. First value is error in difference in objective value, then pointwise error L1, L2 and LInf norm.

Intended to be called via a wrapper to ensure the discretized analytical solution and the numerical solution match the mesh.

compute_errors(
    anasol::Vector{Float64},
    data::PLaplaceData
) -> Vector{Float64}

Same as previous compute_errors(...). However takes discretized analytical solution and PLaplaceData(@ref) as arguments.

PLaplace.write_error_header — Function

write_error_header(filename::String; guarded) -> Int64

Clears and writes a header for an error log to the given file.

PLaplace.check_error_header — Function

check_error_header(filename::String) -> Bool

Checks if given file exists and alredy contains an error header.

PLaplace.write_error — Function

write_error(
    file_name::String,
    data::PLaplaceData,
    error::Vector{Float64}
) -> Int64

Writes a given error along other information corresponding to the header to a given file.

PLaplace.read_error — Function

read_error(fileName::String) -> ErrorData

Returns ErrorData when handed an file generated by PLaplace.

Base.append! — Method

append!(
    A::ErrorData,
    B::ErrorData
) -> Vector{Vector{Float64}}

Appends to objects of type ErrorData by appending all elements of B to the elements in A.

Other Utilities

PLaplace.xpnorm — Function

xpnorm(
    p::Float64,
    dv::AbstractDict{Tuple{Int64, Int64}, AbstractVector{Float64}},
    w::AbstractVector{Float64}
) -> Float64

Returns $\Vert f \Vert^p_{X_p(\Omega)} = \int_{\Omega} \Vert ∇ f(x) \Vert_2^p \;\mathrm{d}x$ in FEM representaion, i.e.

\[\Vert f \Vert^p_{X_p(T_{h_\Omega})} = \sum\limits_{i=1}^m \omega_i \left(\sum\limits_{j=1}^d \sum\limits_{r=1}^{d^\prime} [D^{(j,r)} v]_i^2 \right)^{\frac{p}{2}}.\]

Parameters can either be given as a function and a mesh or with the function as FEM coefficient vector and a derivative tensor.

Mandatory Arguments

p::Float64: Parameter for the norm.

dv::AbstractDict{Tuple{Int64,Int64},AbstractVector{Float64}}: Discrete derivative of function v given at the quadrature nodes of the elements in $T_{h_\Omega}$.

w::AbstractVector{Float64}: Vector of weights corresponding to the quadrature nodes.

xpnorm(
    p::Float64,
    v::AbstractVector{Float64},
    D::AbstractDict{Tuple{Int64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}},
    w::AbstractVector{Float64}
) -> Float64

Does the same as previous xpnorm(...), but takes coefficient vector and derivate tensor as arguments instead of derivative vector.

Requires new mandatory arguments

v::AbstractVector{Float64}: Function discretely evaluated at the nodes of a mesh.

D::AbstractDict{Tuple{Int64,Int64},SparseMatrixCSC{Float64,Int64}}: Discrete derivative tensor corresponding to the same nodes as v.

which replace

dv::AbstractDict{Tuple{Int64,Int64},AbstractVector{Float64}}

xpnorm(
    p::Float64,
    f::Function,
    mesh::Mesh;
    qdim::Int64 = 1
) -> Float64

Does the same as previous xpnorm(...), but takes a continous function and a mesh as arguments instead of a discrete derivative (tensor) and coefficient vector.

Requires new mandatory arguments

f::Function: Analytic description of the function to be evaluated.

mesh::Mesh: FEM mesh corresponding to $T_{h_\Omega}$.

which replace

dv::AbstractDict{Tuple{Int64,Int64},AbstractVector{Float64}}

PLaplace.compute_lipschitzconstant_boundary — Function

compute_lipschitzconstant_boundary(
    gh::Array{Float64,1},
    coords::Array{Array{Float64,1},1},
    boundaryNodes::Set{Int64};
    qdim::Int64 = 1,
    pnorm::Real = 2
)

Returns Lipschitz constant of function g given as FEM coefficient vector on a boundary of the mesh.

Mandatory Arguments

gh::Array{Float64,1}: Function discretely evaluated on a mesh.

coords::Array{Array{Float64,1},1}: Array of coordinates of the nodes gh was evaluated on. In particular contains the coordinates of the boundaryNodes.

boundaryNodes::Set{Int64}: Set of nodes, i.e. entries in coords that belong to the boundary that shall be evaluated. Therefore also selects entries of gh. Potentially shifted by qdim.

Keyword Arguments

qdim::Int64 = 1: Number of components of the function gh.

pnorm::Real = 2: Parameter of p-norm used to measure internal distances.