Internal Documentation

Documentation for PLaplace.jl's internal types and methods.

See the Public page of the library for the documentation of the public interface.

Contents

Internal Documentation

Contents

Utility

Linear Systems

Static Data

Barriers

Tracking

Iteration Data

Log Base

Log Data

Algorithm Data

Path-Following

p-Laplace Data

Algorithm

Problem

Index

Utility

PLaplace.emptyfunction — Method

emptyfunction(args...)

Dummy function for do nothing.

Linear Systems

PLaplace.factorization_none — Method

factorization_none(
    A::SparseArrays.SparseMatrixCSC{Float64, Int64}
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

Wrapper for not computing a factorization of a matrix A, but keeping the matrix itself. Can be used as do nothing factorization for iterative linear solvers.

PLaplace.preconditioner_amg — Method

preconditioner_amg(
    A::SparseArrays.AbstractSparseMatrix{Float64, Ti} where Ti
) -> Preconditioners.AMGPreconditioner{Preconditioners.RugeStuben, _A, AlgebraicMultigrid.V} where _A<:AlgebraicMultigrid.MultiLevel

Wrapper for computing an algebraic multi-grid preconditioner for a matrix A.

PLaplace.preconditioner_illt — Method

preconditioner_illt(
    A::SparseArrays.AbstractSparseMatrix{Float64, Ti} where Ti
) -> Preconditioners.CholeskyPreconditioner{_A} where _A

Wrapper for computing an incomplete Cholesky preconditioner for a matrix A.

PLaplace.preconditioner_ilu — Method

preconditioner_ilu(
    A::SparseArrays.SparseMatrixCSC{Float64, Int64}
) -> IncompleteLU.ILUFactorization{Float64, Int64}

Wrapper for computing an incomplete LU preconditioner for a matrix A.

PLaplace.preconditioner_jacobi — Method

preconditioner_jacobi(
    A::SparseArrays.SparseMatrixCSC{Float64, Int64}
) -> Preconditioners.DiagonalPreconditioner{Float64, Vector{Float64}}

Wrapper for computing a Jacobi (/Diagonal) preconditioner for a matrix A.

PLaplace.preconditioner_none — Method

preconditioner_none(A) -> LinearAlgebra.UniformScaling{Bool}

Wrapper for computing a preconditioner for a matrix A or its factorization without effect, i.e. returns identity.

PLaplace.select_linearsolver — Method

select_linearsolver(
    solver::LinearSolver
) -> Tuple{Function, Function}

Returns pair of functions of the actual solving routine and the factorization routine for the given linear solver.

PLaplace.select_preconditioner — Method

select_preconditioner(
    preconditioner::Preconditioner
) -> Function

Returns function for computing the given preconditioner.

PLaplace.solve_bicgstab — Method

solve_bicgstab(
    A::SparseArrays.SparseMatrixCSC{Float64, Int64},
    b::AbstractVector{Float64},
    P
) -> Any

Wrapper for solving linear system with a preconditioned stabilized bi-conjugated gradients method.

PLaplace.solve_cg — Method

solve_cg(
    A::SparseArrays.SparseMatrixCSC{Float64, Int64},
    b::AbstractVector{Float64},
    P
) -> Any

Wrapper for solving linear system with a preconditioned conjugated gradients method.

PLaplace.solve_default — Method

solve_default(A, b::AbstractVector{Float64}, P) -> Any

Wrapper for solving linear system with julias default \ operator. Usually intended to be called not with a matrix A itself but a factorization, e.g. LU or Cholesky.

PLaplace.solve_gmres — Method

solve_gmres(
    A::SparseArrays.SparseMatrixCSC{Float64, Int64},
    b::AbstractVector{Float64},
    P
) -> Any

Wrapper for solving linear system with a preconditioned GMRES.

Static Data

PLaplace.StaticData — Type

mutable struct StaticData

Stores all data that is static throughout iteration of the interior-point method. Is supposed to be generated by corresponding constructor.

Fields

p::Float64: PDE parameter

d::Int8: Domain dimension

qdim::Int64: Image dimension, i.e. number of components

n::Int64: Number of nodes in domain mesh

m::Int64: Number of elements in domain mesh

calculation_nodes::Set{Int64}: Set of actually calculated nodes

nCalc::Int64: Number of calculated nodes

lengthu::Int64: Number of calculated FEM coefficients

c::AbstractVector{Float64}: System vector for convex optimization problem

R::Float64: Upper bound for auxilliary variable

omega::AbstractVector{Float64}: Vector of element weights

D::AbstractDict{Tuple{Int64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}}: Discrete derivative matrix

g::AbstractVector{Float64}: Prolonged boundary

b::AbstractDict{Tuple{Int64, Int64}, AbstractVector{Float64}}: Discrete derivative of the prolonged boundary condition, i.e. Dg

sigma::Int64: Constant for barrier self-concordance order

alpha::Float64: Constant factor for barrier terms

beta::Float64: Algorithm constant given by Nesterov as 1/9

gamma::Float64: Algorithm constant given by Nesterov as 5/36

tolInv::Float64: Inverse tolerance used for main path-following termination

tolFactor::Float64: Accuracy to tolerance factor

maxIter::Int64: Maximal number of iterations in path-following schemes

kappa::Float64: Initial stepsize for long and adaptive path-following

kappa_updates::Array{Int64, N} where N: Ranges of slow steps for update of kappa in adaptive path-following

kappa_powers::Array{Float64, N} where N: Powers in the update that is applied to kappa in the relevant ranges

maxIterationsBacktracking::Int64: Maximal number of iterations for backtracking in long and adaptive path-following

decrementFactorBacktracking::Float64: Decrement factor for backtracking in long and adaptive path-following

solveLS::Function: Function pointer on solving routine

factorize::Function: Funciton pointer on factorization routine

computePreconditioner::Function: Function pointer on preconditioner

PLaplace.StaticData — Method

StaticData(
    p::Float64,
    mesh::Mesh,
    dirichlet_nodes::Set{Int64},
    fixed_nodes::Union{Missing, Set{Int64}},
    neumann_elements::Union{Missing, Set{Int64}},
    f::Union{Missing, AbstractVector{Float64}},
    g::AbstractVector{Float64},
    h::Union{Missing, AbstractVector{Float64}},
    boundary_prolongation::Union{Missing, AbstractVector{Float64}},
    qdim::Int64,
    eps::Float64,
    maxIter::Int64,
    kappa::Float64,
    kappa_updates::Array{Int64, N} where N,
    kappa_powers::Array{Float64, N} where N,
    maxIterationsBacktracking::Int64,
    decrementFactorBacktracking::Float64,
    solver::LinearSolver,
    preconditioner::Preconditioner,
    useHarmonicProlongation::Bool
) -> PLaplace.StaticData

Constructor for a StaticData object. In particular computes boundary prolongation if required and its derviative, upper and lower bounds on the auxiliary variable as well as the system vector.

PLaplace.compute_sigma — Method

compute_sigma(p::Float64) -> Int64

Returns factor σ for the barrier function parameter depending on p.

PLaplace.compute_systemvector — Method

compute_systemvector(
    mesh::Mesh,
    p::Float64,
    omega::AbstractVector{Float64},
    dirichlet_nodes::Set{Int64},
    has_neumannboundary::Bool,
    neumann_elements::Union{Missing, Set{Int64}},
    h::Union{Missing, AbstractVector{Float64}},
    has_sourceterm::Bool,
    f::Union{Missing, AbstractVector{Float64}},
    qdim::Int64
) -> Any

Computes system vector c for a p-Poisson problem to solve with a barrier method.

PLaplace.compute_upperbound — Method

compute_upperbound(
    mesh::Mesh,
    p::Float64,
    b::AbstractDict{Tuple{Int64, Int64}, AbstractVector{Float64}},
    omega::AbstractVector{Float64},
    has_neumannboundary::Bool,
    neumann_elements::Union{Missing, Set{Int64}},
    h::Union{Missing, AbstractVector{Float64}},
    has_sourceterm::Bool,
    f::Union{Missing, AbstractVector{Float64}},
    qdim::Int64
) -> Float64

Computes constant upper bound R on the auxiliary variable.

Barriers

PLaplace.BarrierFunction — Type

BarrierFunction

Object holdigng the relevant functions for evaluating a barrier function. Includes gradient and hessian as well as a specific initial guess.

Is not intended to be initialized directly but via sub-constructors corresponding to specific barrier files. See their documentations belov to identify the behaviour of specific barrier implemtations.

Fields

isadmissible::Function: Returns boolean value wether $x$ is in the corresponding admissible set.

value::Function: Returns float value of the barrier function at $x$.

gradient::Function: Returns vector-valued gradient of the barrier function at $x$.

hessian::Function: Returns matrix-valued Hessian of the barrier function at $x$.

initialguess::Function: Returns inital vector guess for starting an auxilliary path-following.

PLaplace.BarrierFunction — Method

BarrierFunction() -> PLaplace.BarrierFunction

Wrapping constructor for the selection of a barrier function.

For the specific behaviour of a barrier see the documentation of the sub-constructors below.

PLaplace.FiniteBarrier — Method

FiniteBarrier() -> PLaplace.BarrierFunction

Returns BarrierFunction object corresponding to default barrier function for finite p.

In this setting the admissible set is given by

\[ \mathcal{Q}_p = \left\{ (u,s) \in \mathbb{R}^n \times \mathbb{R}^m \, : \, s_i \geq \left(\sum_{j=1}^{d} \sum_{r=1}^{d'} [D^{(j,r)}(u+g)]_i^2 \right)^{\frac{p}{2}} \land \omega_i s_i \leq R \right\}\]

Defining the variables

\[ z_i(u,s) = s^{\frac{2}{p}} - \sum\limits_{j=1}^d \sum\limits_{r=1}^d [\underbrace{D^{(j,r)}(u+g)}_{=: y^{(j,r)}(u)}]_i^2 \quad \text{and} \quad \tau_i(s) = R -\omega_i s_i\]

it can be equipped with the self-concordant barrier

\[ F(x) = F(u,s) = -\sum\limits_{i=1}^{m} \log(z_i) -\sum\limits_{i=1}^{m} \log(s_i) -\sum\limits_{i=1}^{m} \log(\tau_i).\]

The derivatives

\[ F'(x) = \begin{bmatrix} F_u \\ F_s \end{bmatrix} \text{ and } F''(x) = \begin{bmatrix} F_{uu} & F_{us} \\ F_{us}^{\intercal} & F_{ss} \end{bmatrix}\]

are then given by

\[ \begin{aligned} F_{u} &=2 \sum_{j=1}^{d} \sum_{r=1}^{d'} [D^{(j,r)}]^{\intercal} \frac{y^{(j,r)}}{z} \\ F_{s} &= -\frac{2}{p}\frac{1}{z}s^{\frac{2}{p}-1} + \frac{\omega}{\tau}\\ F_{uu} &= 2 \sum_{j=1}^{d} \sum_{r=1}^{d'} [D^{(j,r)}]^{\intercal} Z^{-1} D^{(j,r)} \\ &\quad+ 4 \sum_{j_1=1}^{d} \sum_{r_1=1}^{d'}\sum_{j_2=1}^{d} \sum_{r_2=1}^{d'} (Y^{(j_1,r_1)}D^{(j_1,r_1)})^{\intercal} Z^{-1} D^{(j)} (Y^{(j_2,r_2)}D^{(j_2,r_2)}),\\ F_{us} &= -\frac{4}{p}\sum_{j=1}^{d} \sum_{r=1}^{d'} (Y^{(j,r)}D^{(j,r)})^{\intercal} Z^{-2} S^{\frac{2}{p}-1},\\ F_{ss} &= -\frac{2}{p} \left(\frac{2}{p} - 1 \right) Z^{-1} S^{\frac{2}{p}-2} + \frac{4}{p^2} Z^{-2} S^{\frac{4}{p}-2} + W^2 T^{-2},\\ \text{where } S &= \mathrm{diag}(s),\, W = \mathrm{diag}(\omega),\, Y = \mathrm{diag}(y),\, Z = \mathrm{diag}(z) \text{ and } T = \mathrm{diag}(\tau). \end{aligned}\]

An initial guess inside $\mathcal{Q}_{\infty}$ is given by

\[ \hat{x} = \begin{bmatrix} 0 \\ \hat{s} \end{bmatrix} \in \mathbb{R}^{n \times m}\]

where

\[ \hat{s} = 1 + \left(\sum_{j=1}^{d} \sum_{r=1}^{d'} [D^{(j,r)} g]_i^2 \right)^{\frac{p}{2}} \in \mathbb{R}^m.\]

Tracking

PLaplace.AssemblyTracker — Type

mutable struct AssemblyTracker

Tracks if factorization or preconditioner got changed during an assembly of the IterationData.

Also tracks condition of system of system matrix, i.e. the barrier Hessian. Has to be tracked via this tool because after the assembly the matrix is usually only available as factorization.

Fields

trackcondition::Bool: Flag if conditino is tracked.

conditionnumber::Union{Missing, Float64}: Condition number of system matrix.

singularity::Bool: Flag if hessian is singular.

factorization::Bool: Flag if factorization got changed.

preconditioner::Bool: Flag if precondtioner got changed.

PLaplace.AssemblyTracker — Method

AssemblyTracker(;trackcondition::Bool = false)

Default constructor for AssemblyTracker.

PLaplace.DescentTracker — Type

mutable struct DescentTracker

Object used to track the result of the application of a descent step. Usually done by a backtracking line search, it contains the required iterations, i.e. updates of the scaling parameter, the final scaling parameter and the required time.

Fields

i::Int64: Number of iterations.

val::Float64: Final scaling value

time::Union{Missing, Float64}: Required time

PLaplace.DescentTracker — Method

DescentTracker(
    i::Int64,
    val::Float64
) -> PLaplace.DescentTracker

Constructor for creating a DescentTracker object with given iteration count and scaling parameter.

PLaplace.DescentTracker — Method

DescentTracker() -> PLaplace.DescentTracker

Constructor for creating a default DescentTracker object.

PLaplace.changed_factorization! — Method

changed_factorization!(
    tracker::PLaplace.AssemblyTracker
) -> Bool

Sets flag of an AssemblyTracker that the factorization got changed.

PLaplace.changed_preconditioner! — Method

changed_preconditioner!(
    tracker::PLaplace.AssemblyTracker
) -> Bool

Sets flag of an AssemblyTracker that the preconditioner got changed.

PLaplace.hessian_singular! — Method

hessian_singular!(tracker::PLaplace.AssemblyTracker) -> Bool

Sets flag of an AssemblyTracker that the hessian is singular and no factorization was computed.

PLaplace.reset! — Method

reset!(tracker::PLaplace.AssemblyTracker) -> Bool

Resets an AssemblyTracker to its default constructor values. Does not change outside flag if condition number is tracked.

PLaplace.reset! — Method

reset!(tracker::PLaplace.DescentTracker) -> Missing

Resets a DescentTracker to its default constructor values.

PLaplace.set! — Method

set!(
    tracker::PLaplace.DescentTracker,
    i::Int64,
    val::Float64
) -> Float64

Sets iteration count and final scaling value to a DescentTracker.

Iteration Data

PLaplace.IterationData — Type

mutable struct IterationData

Stores all data that changes during the iteration of the interior-point method. Especially computes gradient and hessian of the barrier function in each iteration.

PLaplace.apply_descent! — Method

apply_descent!(I::IterationData, S::StaticData; useBacktracking=true)

Performs backtracking with descent direction I.dx required for long and adaptive stepsizes

PLaplace.assemble! — Method

assemble!(I::IterationData, S::StaticData)

Assembles iteration values dependend on the present x and static data.

PLaplace.set! — Method

set!(I::IterationData, x::AbstractVector{Float64}, S::StaticData)

Sets iteration vector and reassembles dependencies, i.e. gradient and hessian.

PLaplace.starnorm — Method

starnorm(
    v::AbstractVector{Float64},
    w::AbstractVector{Float64}
) -> Any

Same as other starnorm(...), but with $w = [F''(x)]^{-1} v$ precomputed.

PLaplace.starnorm — Method

starnorm(
    v::AbstractVector{Float64},
    I::PLaplace.IterationData,
    solveLS::Function
) -> Any

Computes norm induced by the barrier $\Vert v \Vert^*_x = \sqrt{v' [F''(x)]^{-1} v}$.

Log Base

PLaplace.cpad — Function

cpad(
    s::Union{AbstractChar, AbstractString},
    n::Integer
) -> String
cpad(
    s::Union{AbstractChar, AbstractString},
    n::Integer,
    p::Union{AbstractChar, AbstractString}
) -> String

Reimplemtation of deprecated function cpad based on current implementation of lpad and rpad.

PLaplace.log_print_change_factorization — Method

log_print_change_factorization(out::Function) -> Any

Prints setup updated designed to fit the log to the given output.

PLaplace.log_print_change_preconditioner — Method

log_print_change_preconditioner(out::Function) -> Any

Prints setup updated designed to fit the log to the given output.

PLaplace.log_print_defaultdata — Method

log_print_defaultdata(
    out::Function,
    mesh::Mesh,
    eps::Float64,
    p::Float64,
    stepsize::String
) -> Any

Prints problem data that had to be specified for the algorithm to the given output.

PLaplace.log_print_header — Method

log_print_header(out::Function) -> Any

Prints block header for the log to the given output.

PLaplace.log_print_iteration — Method

log_print_iteration(
    out::Function,
    itr::Union{Missing, Int64},
    offset::Int64,
    prefix::String,
    type::Union{Missing, String},
    lastaccept::Union{Missing, Int64},
    kappa::Union{Missing, Float64},
    accnorm::Union{Missing, Float64},
    searchitr::Union{Missing, Int64},
    searchval::Union{Missing, Float64},
    crit::Union{Missing, Float64},
    t::Union{Missing, Float64},
    bound::Float64
) -> Any

Prints one iteration for the log to the given output. In particular contains all relevant information and format parsing.

PLaplace.log_print_line — Method

log_print_line(out::Function) -> Any

Prints a vertical line of width corresponding to the log to the given output.

PLaplace.log_print_setup — Method

log_print_setup(out::Function, stage::Int64) -> Any

Prints setup updated designed to fit the log to the given output.

PLaplace.log_print_spacing — Method

log_print_spacing(out::Function) -> Any

Prints spacing between log blocks to the given output.

PLaplace.log_print_start — Method

log_print_start(out::Function) -> Any

Prints start of log to the given output.

PLaplace.log_print_statistics — Method

log_print_statistics(
    out::Function,
    tsetup::Union{Missing, Float64},
    taux::Union{Missing, Float64},
    tmain::Union{Missing, Float64},
    Naux::Union{Missing, Int64},
    Nmain::Union{Missing, Int64},
    eps::Float64,
    msg::String
) -> Any

Prints statistics of the full solution process to the given output. In particular contains required iterations and time as well as termination messages.

PLaplace.log_print_text — Method

log_print_text(out::Function, text::String) -> Any

Prints a text shorter than the width of the log right aligned with frame to the given output.

PLaplace.log_print_text_center — Method

log_print_text_center(out::Function, text::String) -> Any

Prints a text shorter than the width of the log center aligned with frame to the given output.

PLaplace.log_write_header — Method

log_write_header(filename::String) -> Int64

Clears and writes a header for an stepwise log to the given file.

PLaplace.log_write_iteration — Method

log_write_iteration(
    filename::String,
    itr::Union{Missing, Int64},
    offset::Int64,
    prefix::String,
    type::Union{Missing, String},
    lastaccept::Union{Missing, Int64},
    kappa::Union{Missing, Float64},
    accnorm::Union{Missing, Float64},
    searchitr::Union{Missing, Int64},
    searchval::Union{Missing, Float64},
    crit::Union{Missing, Float64},
    t::Union{Missing, Float64},
    bound::Float64,
    condition::Union{Missing, Float64}
) -> Int64

Writes one iteration for an stepwise log to the given file. In particular contains all relevant information and format parsing.

Log Data

PLaplace.LogData — Type

mutable struct LogData

Structure to hold general information considering the log. In particular if the algorithm is verbose and if a log file is generated.

Fields

isverbose::Bool: Flag indicating if output is verbose.

out::Function: Output function is print() if verbose otherwise does nothing.

logobjective::Bool: Flag for logging the objective to file during the iteration.

trackcondition::Bool: Flag for logging the condition of the linear system to file during the iteration.

file::Union{Missing, String}: File name for writing log to file.

PLaplace.LogData — Method

LogData(
    verbose::Bool,
    filename::Union{Missing, String},
    logcondition::Bool
) -> PLaplace.LogData

Constructor for LogData with data.

PLaplace.LogData — Method

LogData() -> PLaplace.LogData

Constructor for LogData with default values.

PLaplace.log_change_factorization — Method

log_change_factorization(data::PLaplace.LogData) -> Any

Handles change of factorization in the log for output stream. Is not relevant for file.

PLaplace.log_change_preconditioner — Method

log_change_preconditioner(data::PLaplace.LogData) -> Any

Handles change of preconditioner in the log for output stream. Is not relevant for file.

PLaplace.log_defaultdata — Method

log_defaultdata(
    data::PLaplace.LogData,
    mesh::Mesh,
    eps::Float64,
    p::Float64,
    stepsize::String
) -> Any

Handles default data in the log for output stream. Is not relevant for file.

PLaplace.log_footer — Method

log_footer(data::PLaplace.LogData) -> Any

Handles footer in the log for output stream. Is not relevant for file.

PLaplace.log_header — Method

log_header(data::PLaplace.LogData) -> Any

Handles header in the log for output stream. Is not relevant for file.

PLaplace.log_iteration — Method

log_iteration(
    data::PLaplace.LogData,
    itr::Union{Missing, Int64},
    offset::Int64,
    prefix::String,
    type::Union{Missing, String},
    lastaccept::Union{Missing, Int64},
    kappa::Union{Missing, Float64},
    accnorm::Union{Missing, Float64},
    searchitr::Union{Missing, Int64},
    searchval::Union{Missing, Float64},
    condition::Union{Missing, Float64},
    crit::Union{Missing, Float64},
    t::Union{Missing, Float64},
    bound::Float64
) -> Any

Handles an iteration in the log for output stream and file generation.

PLaplace.log_line — Method

log_line(data::PLaplace.LogData) -> Any

Handles horizontal line in the log for output stream. Is not relevant for file.

PLaplace.log_setup — Method

log_setup(data::PLaplace.LogData, stage::Int64) -> Any

Handles setup in the log for output stream. Is not relevant for file.

PLaplace.log_start — Method

log_start(data::PLaplace.LogData) -> Any

Handles start block in the log for output stream. Is not relevant for file.

PLaplace.log_statistics — Method

log_statistics(
    data::PLaplace.LogData,
    tsetup::Union{Missing, Float64},
    taux::Union{Missing, Float64},
    tmain::Union{Missing, Float64},
    Naux::Union{Missing, Int64},
    Nmain::Union{Missing, Int64},
    eps::Float64,
    msg::String
) -> Any

Handles statistics in the log for output stream. Is not relevant for file.

PLaplace.log_text — Method

log_text(data::PLaplace.LogData, msg::String) -> Any

Handles text line in the log for output stream. Is not relevant for file.

Algorithm Data

PLaplace.AlgorithmData — Type

mutable struct AlgorithmData

Structure to store statistics to the algorithm during the runtime. The information is supposed to be exported to PLaplaceData for external usage.

Fields

eps::Float64: Obtained accuracy.

Naux::Union{Missing, Int64}: Required iterations for the auxilliary path-following.

Nmain::Union{Missing, Int64}: Required iterations for the main path-following.

tsetup::Union{Missing, Float64}: Required time for the setup.

taux::Union{Missing, Float64}: Required time for the auxilliary path-following.

tmain::Union{Missing, Float64}: Required time for the main path-following.

solution::Union{Missing, Vector{Float64}}: Solution vector if iteration converged.

msg::String: Notifications from the iteration.

PLaplace.AlgorithmData — Method

AlgorithmData() -> PLaplace.AlgorithmData

Constructor for AlgorithmData with empty values.

PLaplace.handle_accuracy! — Method

handle_accuracy!(
    data::PLaplace.AlgorithmData,
    eps::Float64
) -> Float64

Handling the termination of a main path-following because of a singular system matrix. In particular stores the resulting (reversly computed) obtained accuracy.

PLaplace.handle_assembly! — Method

handle_assembly!(
    data::PLaplace.AlgorithmData,
    log::PLaplace.LogData,
    tracker::PLaplace.AssemblyTracker,
    phase::String,
    iteration::Int64
) -> Any

Handling the assembly of barrier terms. In particular stores message if solver or preconditioner got changed.

PLaplace.handle_auxfail! — Method

handle_auxfail!(data::PLaplace.AlgorithmData) -> String

Handling the termination of the algorithm because the final update step in the auxilliary path-following failed because of a singular system matrix.

PLaplace.handle_kappavanish! — Method

handle_kappavanish!(
    data::PLaplace.AlgorithmData,
    phase::String,
    iteration::Int64
) -> String

Handling the termination of the algorithm because the stepsize update parameter κ in a path-following with adaptive stepping got numerically too small. Should theoretically not occur, so this usually indicates an infeasible problem.

PLaplace.handle_maxiterations! — Method

handle_maxiterations!(
    data::PLaplace.AlgorithmData,
    phase::String,
    iteration::Int64
) -> String

Handling the termination of the algorithm because the maximum number of iterations during a phase was reached.

PLaplace.log_statistics — Method

log_statistics(
    A::PLaplace.AlgorithmData,
    L::PLaplace.LogData
) -> Any

Wrapper for passing statistics to the log handling.

Path-Following

PLaplace.pathfollowing_auxilliary_adaptive! — Method

pathfollowing_auxilliary_adaptive!(
    I::PLaplace.IterationData,
    A::PLaplace.AlgorithmData,
    S::PLaplace.StaticData,
    L::PLaplace.LogData
) -> Int64

Executes auxilliary path-following with adaptive stepsize.

The iteration is performed on the IterationData, which will later store the final iterate as well as all the corresponding barrier terms. The number of required iterations and potential messages will be stored in AlgorithmData. If LogData is verbose, data per iteration will be written to the output stream. Further, if a file is provided, the log will be also exported to that file.

PLaplace.pathfollowing_auxilliary_long! — Method

pathfollowing_auxilliary_long!(
    I::PLaplace.IterationData,
    A::PLaplace.AlgorithmData,
    S::PLaplace.StaticData,
    L::PLaplace.LogData
) -> Int64

Executes auxilliary path-following with long stepsize.

The iteration is performed on the IterationData, which will later store the final iterate as well as all the corresponding barrier terms. The number of required iterations and potential messages will be stored in AlgorithmData. If LogData is verbose, data per iteration will be written to the output stream. Further, if a file is provided, the log will be also exported to that file.

PLaplace.pathfollowing_auxilliary_short! — Method

pathfollowing_auxilliary_short!(
    I::PLaplace.IterationData,
    A::PLaplace.AlgorithmData,
    S::PLaplace.StaticData,
    L::PLaplace.LogData
) -> Int64

Executes auxilliary path-following with short stepsize.

The iteration is performed on the IterationData, which will later store the final iterate as well as all the corresponding barrier terms. The number of required iterations and potential messages will be stored in AlgorithmData. If LogData is verbose, data per iteration will be written to the output stream. Further, if a file is provided, the log will be also exported to that file.

PLaplace.pathfollowing_main_adaptive! — Method

pathfollowing_main_adaptive!(
    I::PLaplace.IterationData,
    A::PLaplace.AlgorithmData,
    S::PLaplace.StaticData,
    L::PLaplace.LogData
) -> Int64

Executes main path-following with adaptive stepsize.

The iteration is performed on the IterationData, which will later store the final result. If the iteration did not converge, the last iterate will still be provided as a result and the obtained accuracy stored in AlgorithmData. The number of required iterations and potential messages will be stored in AlgorithmData in any case. If LogData is verbose, data per iteration will be written to the output stream. Further, if a file is provided, the log will be also exported to that file.

PLaplace.pathfollowing_main_long! — Method

pathfollowing_main_long!(
    I::PLaplace.IterationData,
    A::PLaplace.AlgorithmData,
    S::PLaplace.StaticData,
    L::PLaplace.LogData
) -> Int64

Executes main path-following with long stepsize.

The iteration is performed on the IterationData, which will later store the final result. If the iteration did not converge, the last iterate will still be provided as a result and the obtained accuracy stored in AlgorithmData. The number of required iterations and potential messages will be stored in AlgorithmData in any case. If LogData is verbose, data per iteration will be written to the output stream. Further, if a file is provided, the log will be also exported to that file.

PLaplace.pathfollowing_main_short! — Method

pathfollowing_main_short!(
    I::PLaplace.IterationData,
    A::PLaplace.AlgorithmData,
    S::PLaplace.StaticData,
    L::PLaplace.LogData
) -> Int64

Executes main path-following with short stepsize.

The iteration is performed on the IterationData, which will later store the final result. If the iteration did not converge, the last iterate will still be provided as a result and the obtained accuracy stored in AlgorithmData. The number of required iterations and potential messages will be stored in AlgorithmData in any case. If LogData is verbose, data per iteration will be written to the output stream. Further, if a file is provided, the log will be also exported to that file.

PLaplace.select_pathfollowing — Method

select_pathfollowing(stepsize::Stepsize) -> Tuple{Function, Function}

Returns functions for auxilliary and main pathfollowing dependend on the stepsize.

p-Laplace Data

PLaplace.add_algorithmdata! — Method

add_algorithmdata!(
    data::PLaplaceData,
    algorithm::PLaplace.AlgorithmData
) -> Union{Nothing, Vector{Float64}}

Adds required values from the iteration data to the output data, especially the result.

PLaplace.add_staticdata! — Method

add_staticdata!(
    data::PLaplaceData,
    S::PLaplace.StaticData
) -> Set{Int64}

Adds required values from the static data to the output data. In particular the prolonged boundary and which nodes will not be fixed during the computation.

PLaplace.assemble_result! — Method

assemble_result!(
    data::PLaplaceData
) -> Union{Nothing, Vector{Float64}}

Assembles combined result vector v = u + g for truncated u vector u_I.

Algorithm

PLaplace.check_parameters — Method

check_parameters(
    p::Float64,
    qdim::Int64,
    eps::Float64,
    kappa::Float64,
    kappa_updates::Array{Int64, N} where N,
    kappa_powers::Array{Float64, N} where N
)

Throws error if given parameters are not in the designated range.

Problem

PLaplace.compute_plaplace_term — Method

compute_plaplace_term(
    u::AbstractVector{Float64},
    p::Float64,
    mesh::Mesh,
    qdim::Int64
) -> Float64

Computes value of characteristic derivative term in p-Laplace functional evaluated at u.

PLaplace.compute_sources — Method

compute_sources(
    u::AbstractVector{Float64},
    mesh::Mesh,
    neumann_boundary::Union{Missing, Set{Int64}, Set{Boundary}},
    h::Union{Missing, Function, AbstractVector{Float64}},
    f::Union{Missing, Function, AbstractVector{Float64}},
    qdim::Int64
) -> Float64

Computes value of the required boundary and volume source terms for the p-Laplace functional evaluated at u.

Index

PLaplace.AlgorithmData

PLaplace.AlgorithmData

PLaplace.AssemblyTracker

PLaplace.AssemblyTracker

PLaplace.BarrierFunction

PLaplace.BarrierFunction

PLaplace.DescentTracker

PLaplace.DescentTracker

PLaplace.DescentTracker

PLaplace.IterationData

PLaplace.LogData

PLaplace.LogData

PLaplace.LogData

PLaplace.StaticData

PLaplace.StaticData

PLaplace.FiniteBarrier

PLaplace.add_algorithmdata!

PLaplace.add_staticdata!

PLaplace.apply_descent!

PLaplace.assemble!

PLaplace.assemble_result!

PLaplace.changed_factorization!

PLaplace.changed_preconditioner!

PLaplace.check_parameters

PLaplace.compute_plaplace_term

PLaplace.compute_sigma

PLaplace.compute_sources

PLaplace.compute_systemvector

PLaplace.compute_upperbound

PLaplace.cpad

PLaplace.emptyfunction

PLaplace.factorization_none

PLaplace.handle_accuracy!

PLaplace.handle_assembly!

PLaplace.handle_auxfail!

PLaplace.handle_kappavanish!

PLaplace.handle_maxiterations!

PLaplace.hessian_singular!

PLaplace.log_change_factorization

PLaplace.log_change_preconditioner

PLaplace.log_defaultdata

PLaplace.log_footer

PLaplace.log_header

PLaplace.log_iteration

PLaplace.log_line

PLaplace.log_print_change_factorization

PLaplace.log_print_change_preconditioner

PLaplace.log_print_defaultdata

PLaplace.log_print_header

PLaplace.log_print_iteration

PLaplace.log_print_line

PLaplace.log_print_setup

PLaplace.log_print_spacing

PLaplace.log_print_start

PLaplace.log_print_statistics

PLaplace.log_print_text

PLaplace.log_print_text_center

PLaplace.log_setup

PLaplace.log_start

PLaplace.log_statistics

PLaplace.log_statistics

PLaplace.log_text

PLaplace.log_write_header

PLaplace.log_write_iteration

PLaplace.pathfollowing_auxilliary_adaptive!

PLaplace.pathfollowing_auxilliary_long!

PLaplace.pathfollowing_auxilliary_short!

PLaplace.pathfollowing_main_adaptive!

PLaplace.pathfollowing_main_long!

PLaplace.pathfollowing_main_short!

PLaplace.preconditioner_amg

PLaplace.preconditioner_illt

PLaplace.preconditioner_ilu

PLaplace.preconditioner_jacobi

PLaplace.preconditioner_none

PLaplace.reset!

PLaplace.reset!

PLaplace.select_linearsolver

PLaplace.select_pathfollowing

PLaplace.select_preconditioner

PLaplace.set!

PLaplace.set!

PLaplace.solve_bicgstab

PLaplace.solve_cg

PLaplace.solve_default

PLaplace.solve_gmres

PLaplace.starnorm

PLaplace.starnorm